research/alchemist-strategic-fusion/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alchemist-strategic-fusion"
version = "0.1.0"
description = "Cross-attention fusion: chess/go strategic state ↔ music token sequences (research)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "transformers>=4.36.0",
  "python-chess>=1.999",
  "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pretty-midi>=0.2.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
